Added Hotel Room Collection Model and Singleton Metaclass
Since there is only one set of data, it makes sense that the data be
loaded once and re-used by all requests to the system, so keeping the
database in a Singleton pattern is efficient. The Hotel Room Collection
is in effect the database for this system. HotelRoomCollection supports
querying for a list of rooms by city, returning None if the city is not
found. The lists of hotel rooms are grouped by city using a dictionary,
for O(1) lookups for these lists. Sorting the lists by price is done
for each request for a sorted list; this can definitely be optimized
(cache the results? pre-sort the list?)

Moved the CSV to a static folder and created test files for testing.